refactor: add azure-functions kit and reuse code that run custom script in vm
