+++ packages/toolkit/azure-functions-kit/CHANGELOG.md
@@ -1,0 +1,8 @@
+# Change Log
+
+## 0.1.0
+
+Init commit of the Azure Functions Kit, including:
+
+1. General error handling for Azure Functions
+2. Shared Azure Functions utilities, such as running `getCustomJwtClaims` logic.
