+++ packages/toolkit/azure-functions-kit/package.json
@@ -1,0 +1,47 @@
+{
+  "name": "@logto/azure-functions-kit",
+  "version": "0.1.0",
+  "author": "Silverhand Inc. <contact@silverhand.io>",
+  "homepage": "https://github.com/logto-io/toolkit#readme",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/logto-io/toolkit.git"
+  },
+  "license": "MPL-2.0",
+  "type": "module",
+  "main": "./lib/index.js",
+  "types": "./lib/index.d.ts",
+  "files": [
+    "lib"
+  ],
+  "scripts": {
+    "precommit": "lint-staged",
+    "build": "rm -rf lib/ && tsc -p tsconfig.build.json",
+    "build:test": "pnpm build",
+    "lint": "eslint --ext .ts src",
+    "lint:report": "pnpm lint --format json --output-file report.json",
+    "prepack": "pnpm build"
+  },
+  "engines": {
+    "node": "^20.9.0"
+  },
+  "devDependencies": {
+    "@azure/functions": "^4.0.0",
+    "@logto/schemas": "workspace:1.15.0",
+    "@silverhand/eslint-config": "6.0.1",
+    "@silverhand/essentials": "^2.9.0",
+    "@silverhand/ts-config": "6.0.0",
+    "@types/node": "^20.9.5",
+    "eslint": "^8.56.0",
+    "lint-staged": "^15.0.0",
+    "prettier": "^3.0.0",
+    "typescript": "^5.3.3"
+  },
+  "eslintConfig": {
+    "extends": "@silverhand"
+  },
+  "prettier": "@silverhand/eslint-config/.prettierrc",
+  "publishConfig": {
+    "access": "public"
+  }
+}
