+++ packages/toolkit/azure-functions-kit/tsconfig.build.json
@@ -1,0 +1,5 @@
+{
+  "extends": "./tsconfig",
+  "include": ["src"],
+  "exclude": ["src/**/*.test.ts"]
+}
