+++ packages/toolkit/azure-functions-kit/tsconfig.json
@@ -1,0 +1,6 @@
+{
+  "extends": "@silverhand/ts-config/tsconfig.base",
+  "compilerOptions": {
+    "outDir": "lib"
+  }
+}
